Build dynamic temporal granularities

Build dynamic temporal granularities

create_gran(.data, gran1 = NULL, hierarchy_tbl = NULL, label = TRUE,
  abbr = TRUE, ...)

Arguments

.data
A tsibble object.

gran1
the granularity to be created. For temporal data, any combination of "second", "minute", "qhour", "hhour", "hour", "day", "week", "fortnight ,"month", "quarter", "semester" or "year" can be chosen in the form of finer to coarser unit. For example, for the granularity hour of the week, value is "hour_week".

hierarchy_tbl
A hierarchy table specifying the hierarchy of units and their relationships

label
Logical. TRUE will display the month as an ordered factor of character string such as "January", "February". FALSE will display the month as an ordered factor such as 1 to 12, where 1 stands for January and 12 for December.

abbr
logical. FALSE will display abbreviated labels

...
Other arguments passed on to individual methods.

Value

A tsibble with an additional column of granularity

Examples

library(gravitas)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(tsibble)

#> 
#> Attaching package: ‘tsibble’

#> The following object is masked from ‘package:dplyr’:
#> 
#>     id

library(ggplot2)

smart_meter <- smart_meter10 %>%
as_tsibble() %>%
ungroup()

# Search for granularities

smart_meter %>%
  search_gran(highest_unit = "month")

#>  [1] "hhour_hour"      "hhour_day"       "hhour_week"      "hhour_fortnight"
#>  [5] "hhour_month"     "hour_day"        "hour_week"       "hour_fortnight" 
#>  [9] "hour_month"      "day_week"        "day_fortnight"   "day_month"      
#> [13] "week_fortnight"  "week_month"      "fortnight_month"

# Screen harmonies from the search list
if (FALSE) {

smart_meter %>%
  harmony(
    ugran = "fortnight",
    filter_out = c("hhour"),
    filter_in = "wknd_wday"
  )

# visualize probability distribution of
# the harmony pair (wknd_wday, hour_day)

smart_meter %>%
  prob_plot(
    gran1 = "wknd_wday",
    gran2 = "hour_day",
    response = "general_supply_kwh",
    plot_type = "quantile",
    quantile_prob = c(0.1, 0.25, 0.5, 0.75, 0.9)
  ) +
  scale_y_sqrt()
}

# Compute granularities for non-temporal data

cricket_tsibble <- cricketdata %>%
  mutate(data_index = row_number()) %>%
  as_tsibble(index = data_index)

hierarchy_model <- tibble::tibble(
  units = c("index", "ball", "over", "inning", "match"),
  convert_fct = c(1, 6, 20, 2, 1)
)
cricket_tsibble %>%
  create_gran(
    "ball_over",
    hierarchy_model
  )

#> # A tsibble: 56,640 x 26 [1]
#>    match_id inning batting_team bowling_team  over  ball batsman non_striker
#>       <dbl>  <dbl> <chr>        <chr>        <dbl> <dbl> <chr>   <chr>      
#>  1        5      1 Royal Chall… Delhi Dared…     1     1 CH Gay… SR Watson  
#>  2        5      1 Royal Chall… Delhi Dared…     1     2 SR Wat… CH Gayle   
#>  3        5      1 Royal Chall… Delhi Dared…     1     3 SR Wat… CH Gayle   
#>  4        5      1 Royal Chall… Delhi Dared…     1     4 SR Wat… CH Gayle   
#>  5        5      1 Royal Chall… Delhi Dared…     1     5 SR Wat… CH Gayle   
#>  6        5      1 Royal Chall… Delhi Dared…     1     6 SR Wat… CH Gayle   
#>  7        5      1 Royal Chall… Delhi Dared…     2     1 SR Wat… CH Gayle   
#>  8        5      1 Royal Chall… Delhi Dared…     2     2 SR Wat… CH Gayle   
#>  9        5      1 Royal Chall… Delhi Dared…     2     3 SR Wat… CH Gayle   
#> 10        5      1 Royal Chall… Delhi Dared…     2     4 CH Gay… SR Watson  
#> # … with 56,630 more rows, and 18 more variables: bowler <chr>,
#> #   is_super_over <dbl>, wide_runs <dbl>, bye_runs <dbl>, legbye_runs <dbl>,
#> #   noball_runs <dbl>, penalty_runs <dbl>, batsman_runs <dbl>,
#> #   extra_runs <dbl>, total_runs <dbl>, player_dismissed <chr>,
#> #   dismissal_kind <chr>, fielder <chr>, cricket_key <chr>,
#> #   ball_per_over <int>, over_key <chr>, data_index <int>, ball_over <fct>

# Validate if given column in the data set
# equals computed granularity

validate_gran(cricket_tsibble,
gran = "over_inning",
hierarchy_tbl = hierarchy_model,
 validate_col = "over")

#> [1] TRUE