Advice summaries for granularities

Recommendations on plot choices, interaction, number of observations and intra or inter facet homogeneity. Important summaries before drawing distribution plots.

# S3 method for gran_advice
print(x, ...)

Arguments

x
An object of class gran_advice

...
other arguments to be passed for appropriate labels.

Value

Print check points before visualizing distribution across bivariate granularities

Examples

library(dplyr)
library(ggplot2)

smart_meter10 %>%
filter(customer_id == "10017936") %>%
  gran_advice(gran1 = "wknd_wday",
            gran2 = "hour_day")

#> The chosen granularities are harmonies 
#>  
#> Recommended plots are: violin lv quantile boxplot 
#>  
#> Number of observations are homogenous across facets 
#>  
#> Number of observations are homogenous within facets 
#>  
#> Cross tabulation of granularities : 
#>  
#> # A tibble: 24 x 3
#>    hour_day Weekday Weekend
#>    <fct>      <dbl>   <dbl>
#>  1 0            910     366
#>  2 1            908     366
#>  3 2            909     366
#>  4 3            910     366
#>  5 4            910     366
#>  6 5            910     366
#>  7 6            909     366
#>  8 7            908     366
#>  9 8            908     366
#> 10 9            908     366
#> # … with 14 more rows