Plotting probability distributions across granularities

Plot probability distribution of univariate series across bivariate temporal granularities.

rank_harmony(
  .data = NULL,
  harmony_tbl = NULL,
  response = NULL,
  prob = seq(0.01, 0.99, 0.01),
  hierarchy_tbl = NULL
)

Arguments

.data
a tsibble.

harmony_tbl
A tibble of harmonies and their levels obtained from the function().

response
response variable.

prob
numeric vector of probabilities with values in [0,1].

hierarchy_tbl
A hierarchy table specifying the hierarchy of units

Value

A tibble of harmonies and their levels ranked ion descending order of average maximum pairwise distance of the harmony pairs.

Examples

library(tsibbledata)
library(ggplot2)
library(tsibble)
library(lvplot)
library(dplyr)
library(gravitas)
library(purrr)
library(magrittr)

#> 
#> Attaching package: ‘magrittr’

#> The following object is masked from ‘package:purrr’:
#> 
#>     set_names

harmonies <- smart_meter10 %>%
 harmony(ugran = "month",
         filter_out = c("hhour", "fortnight"))
.data = smart_meter10
response  = "general_supply_kwh"
harmony_tbl =  harmonies
.data %>% rank_harmony(harmony_tbl = harmonies, response = "general_supply_kwh")

#> # A tibble: 10 x 5
#>    facet_variable x_variable facet_levels x_levels mean_max_variation
#>    <chr>          <chr>             <int>    <int>              <dbl>
#>  1 hour_day       week_month           24        5               0.2 
#>  2 hour_day       day_week             24        7               0.17
#>  3 day_month      day_week             31        7               0.12
#>  4 hour_day       day_month            24       31               0.11
#>  5 day_month      hour_day             31       24               0.06
#>  6 day_week       day_month             7       31               0.03
#>  7 day_week       hour_day              7       24               0.01
#>  8 week_month     hour_day              5       24               0.01
#>  9 week_month     day_week              5        7               0.01
#> 10 day_week       week_month            7        5               0.01

harmony_tbl <- PBS %>% harmony(ugran = "year")
rank_harmony(PBS, harmony_tbl = harmony_tbl, response = "Cost")

#> # A tibble: 10 x 5
#>    facet_variable   x_variable       facet_levels x_levels mean_max_variation
#>    <chr>            <chr>                   <int>    <int>              <dbl>
#>  1 month_semester   semester_year               6        2               0.07
#>  2 quarter_year     month_quarter               4        3               0.04
#>  3 semester_year    quarter_semester            2        2               0.02
#>  4 quarter_semester month_quarter               2        3               0.01
#>  5 semester_year    month_quarter               2        3               0.01
#>  6 semester_year    month_semester              2        6               0.01
#>  7 month_quarter    quarter_semester            3        2               0.01
#>  8 month_quarter    quarter_year                3        4               0.01
#>  9 month_quarter    semester_year               3        2               0.01
#> 10 quarter_semester semester_year               2        2               0.01