Plotting probability distributions across granularities

Plot probability distribution of univariate series across bivariate temporal granularities.

rank_harmony_v0(
  .data = NULL,
  harmony_tbl = NULL,
  response = NULL,
  prob = seq(0.01, 0.99, 0.01),
  dist_distribution = "normal",
  hierarchy_tbl = NULL,
  dist_ordered = TRUE,
  alpha = 0.05
)

Arguments

.data
a tsibble.

harmony_tbl
A tibble of harmonies and their levels obtained from the function().

response
response variable.

prob
numeric vector of probabilities with values in [0,1].

dist_distribution
Underlying distribution of distances. Look at hist_distance()

hierarchy_tbl
A hierarchy table specifying the hierarchy of units

dist_ordered
if levels of the time granularity is ordered.

alpha
significance level

Value

A tibble of harmonies and their levels ranked in descending order of average maximum pairwise distance of the harmony pairs.